+++ src/models/lessonmodel.js
@@ -1,26 +1,26 @@
-import mongoose from 'mongoose';
+import mongoose from "mongoose";
 const { ObjectId } = mongoose.Schema.Types;
 
 const lessonSchema = new mongoose.Schema(
-    {
-        name: {
-            type: String,
-            required: true
-        },
-        description: {
-            type: String,
-            // required: true,
-        },
-        videoId: {
-            type: String,
-            required: true,
-        },
-        courseId: {
-            type: ObjectId,
-            ref: "Course",
-        },
-    },
-    { timestamps: true }
+	{
+		name: {
+			type: String,
+			required: true
+		},
+		description: {
+			type: String
+			// required: true,
+		},
+		videoId: {
+			type: String,
+			required: true
+		},
+		courseId: {
+			type: ObjectId,
+			ref: "Course"
+		}
+	},
+	{ timestamps: true }
 );
-
-export const Lesson = mongoose.model("Lesson", lessonSchema);
+const LessonModel = mongoose.model("Lesson", lessonSchema);
+export default LessonModel;
