+++ src/routes/LessonRouter.js
@@ -1,0 +1,7 @@
+import express from "express";
+import LessonController from "../controllers/lessonController.js";
+const router = express.Router();
+
+router.post("/", LessonController.createLesson);
+
+export default router;
